Added yml file , deleted prune and added it to same file
